Colored Object Tracker
This program applies color masks/filters to a pre-selected image and highlights the objects with pre-selected colors in the image, for me the algorithm makes a binarization of the image leaving the object of interest in white and the rest of the image in black. We also have the calculation of the center of mass of the image and its contour as well as the histogram.